Extract config DSL to its own class, add unit test
